Changed all ushort data types to int. Added DisplayResults and DisplayPlayerInventory methods to UI class. Added option in Validation for StoreItems and RecipeItems to be able to cancel. Moved ChangeRecipe from Recipe to Human Player class for now (allows for direct access to the player's inventory for validation). Added game instantiation to Program class. Added inventoryDifference method to Game class to calculate difference between two different inventories, such as day 1 and day 2, day 1 and day 7, etc. Related, added List of Inventory class to track multiple days of Inventory.

+++ LemonadeStand/Inventory.cs
@@ -8,10 +8,10 @@ namespace LemonadeStand
   public class Inventory
   {
     public double money;
-    public ushort cups;
-    public ushort lemons;
-    public ushort sugar;
-    public ushort ice;
+    public int cups;
+    public int lemons;
+    public int sugar;
+    public int ice;
 
     public Inventory()
     {
